scraper/requirements.txt
selenium
psycopg2-binary
# sbvirtualdisplay
# chromedriver